The rating function is fully implemented through out the app.

+++ lib/models/rating.dart
@@ -20,7 +20,7 @@ class Rating {
   factory Rating.fromMap(Map<String, dynamic> map) {
     return Rating(
       userId: map['userId'] as String,
-      rating: map['rating'] as double,
+      rating: ( map['rating'] as int).toDouble(),
     );
   }
